常规功能添加

+++ MyMusic/搜索界面/NotesTableViewCell.h
@@ -12,7 +12,7 @@
 
 @interface NotesTableViewCell : UITableViewCell
 
-@property (assign, nonatomic) int lineNo;
+@property (assign, nonatomic) long lineNo;
 
 @property (strong, nonatomic) IBOutlet UIImageView *notesView;
